bugfix menores y reorganizacion
