+++ src/app/calculadora/calculadora.component.html
@@ -1,6 +1,6 @@
 <div class="row gx-3" style="height: 100vh; background: linear-gradient(120deg, #f4f4f9, #e8f5e9);">
   <!-- Columna izquierda con tarjetas -->
-  <div class="col-3 d-flex flex-column justify-content-around">
+  <div class="col-3 d-flex justify-content-around align-self-top">
       <app-data-container [title]="titles[activeTab]"
                           [data]="dataSets[activeTab]"></app-data-container>
   </div>
