+++ src/app/herramientass/recolector-de-datos/recolector-de-datos.component.html
