+++ src/app/portada/portada.component.html
@@ -24,8 +24,7 @@
       <li>Ciencias sociales (encuestas, demografía).</li>
     </ul>
     <a class="nav-link"
-       routerLink="/calculadora"
-       routerLinkActive="active"
+       (click)="navegar('/calculadora')"
        style="display: inline-block; text-decoration: none; background-color: #2c3e50; color: white; padding: 10px 20px; border-radius: 5px; font-size: 1rem; text-align: center; margin-top: 20px;">
       Ir a la Calculadora
     </a>
